+++ schemify.php
@@ -2,7 +2,7 @@
 /**
  * Plugin Name: Schemify
  * Plugin URI:  https://stevegrunwell.com/schemify
- * Description: Automatically generate Schema.org JSON+LD markup for WordPress.
+ * Description: Automatically generate Schema.org JSON-LD markup for WordPress.
  * Version:     0.1.0
  * Author:      Steve Grunwell
  * Author URI:  https://stevegrunwell.com